+++ Library/Source/Nano/Internal/3rd Party/schneier.com/blowfish.h
@@ -10,17 +10,17 @@ extern "C" {
 
 
 typedef struct blowfish_context_t_ {
-	unsigned long pbox[256];
-	unsigned long sbox[4][256];
+	UInt32 pbox[256];
+	UInt32 sbox[4][256];
 } blowfish_context_t;
 
-void blowfish_encryptblock (blowfish_context_t *ctx, unsigned long  *hi,  unsigned long *lo);
-void blowfish_decryptblock (blowfish_context_t *ctx, unsigned long  *hi,  unsigned long *lo);
-void blowfish_initiate     (blowfish_context_t *ctx, void *keyparam,      unsigned int  keybytes);
+void blowfish_encryptblock (blowfish_context_t *ctx, UInt32  *hi,  UInt32 *lo);
+void blowfish_decryptblock (blowfish_context_t *ctx, UInt32  *hi,  UInt32 *lo);
+void blowfish_initiate     (blowfish_context_t *ctx, void *keyparam,      UInt32  keybytes);
 void blowfish_clean        (blowfish_context_t *ctx);
 
-extern const unsigned long ORIG_P[18];
-extern const unsigned long ORIG_S[4][256];
+extern const UInt32 ORIG_P[18];
+extern const UInt32 ORIG_S[4][256];
